# api-docs_1.2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Discodeit API 문서",
    "description": "Discodeit 프로젝트의 Swagger API 문서입니다.",
    "version": "1.2"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "로컬 서버"
    }
  ],
...
  "paths": {
...
    "/api/readStatuses": {
      "get": {
        "tags": [
          "ReadStatus"
        ],
        "summary": "User의 Message 읽음 상태 목록 조회",
        "operationId": "findAllByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "조회할 User ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message 읽음 상태 목록 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadStatusDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ReadStatus"
        ],
        "summary": "Message 읽음 상태 생성",
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadStatusCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "이미 읽음 상태가 존재함",
            "content": {
              "*/*": {
                "example": "ReadStatus with userId {userId} and channelId {channelId} already exists"
              }
            }
          },
          "201": {
            "description": "Message 읽음 상태가 성공적으로 생성됨",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStatusDto"
                }
              }
            }
          },
          "404": {
            "description": "Channel 또는 User를 찾을 수 없음",
            "content": {
              "*/*": {
                "example": "Channel | User with id {channelId | userId} not found"
              }
            }
          }
        }
      }
    },
    "/api/messages": {
      "get": {
        "tags": [
          "Message"
        ],
        "summary": "Channel의 Message 목록 조회",
        "operationId": "findAllByChannelId",
        "parameters": [
          {
            "name": "channelId",
            "in": "query",
            "description": "조회할 Channel ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "페이징 커서 정보",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "description": "페이징 정보",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            },
            "example": {
              "size": 50,
              "sort": "createdAt,desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message 목록 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Message"
        ],
        "summary": "Message 생성",
        "operationId": "create_2",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "messageCreateRequest": {
                    "$ref": "#/components/schemas/MessageCreateRequest"
                  },
                  "attachments": {
                    "type": "array",
                    "description": "Message 첨부 파일들",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                },
                "required": [
                  "messageCreateRequest"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message가 성공적으로 생성됨",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          },
          "404": {
            "description": "Channel 또는 User를 찾을 수 없음",
            "content": {
              "*/*": {
                "example": "Channel | Author with id {channelId | authorId} not found"
              }
            }
          }
        }
      }
    },
    "/api/channels/public": {
      "post": {
        "tags": [
          "Channel"
        ],
        "summary": "Public Channel 생성",
        "operationId": "create_3",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicChannelCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Public Channel이 성공적으로 생성됨",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/channels/private": {
      "post": {
        "tags": [
          "Channel"
        ],
        "summary": "Private Channel 생성",
        "operationId": "create_4",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrivateChannelCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Private Channel이 성공적으로 생성됨",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/messages/{messageId}": {
      "delete": {
        "tags": [
          "Message"
        ],
        "summary": "Message 삭제",
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "description": "삭제할 Message ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Message가 성공적으로 삭제됨"
          },
          "404": {
            "description": "Message를 찾을 수 없음",
            "content": {
              "*/*": {
                "example": "Message with id {messageId} not found"
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Message"
        ],
        "summary": "Message 내용 수정",
        "operationId": "update_2",
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "description": "수정할 Message ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Message를 찾을 수 없음",
            "content": {
              "*/*": {
                "example": "Message with id {messageId} not found"
              }
            }
          },
          "200": {
            "description": "Message가 성공적으로 수정됨",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/channels/{channelId}": {
      "delete": {
        "tags": [
          "Channel"
        ],
        "summary": "Channel 삭제",
        "operationId": "delete_2",
        "parameters": [
          {
            "name": "channelId",
            "in": "path",
            "description": "삭제할 Channel ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Channel을 찾을 수 없음",
            "content": {
              "*/*": {
                "example": "Channel with id {channelId} not found"
              }
            }
          },
          "204": {
            "description": "Channel이 성공적으로 삭제됨"
          }
        }
      },
      "patch": {
        "tags": [
          "Channel"
        ],
        "summary": "Channel 정보 수정",
        "operationId": "update_3",
        "parameters": [
          {
            "name": "channelId",
            "in": "path",
            "description": "수정할 Channel ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicChannelUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Channel을 찾을 수 없음",
            "content": {
              "*/*": {
                "example": "Channel with id {channelId} not found"
              }
            }
          },
          "400": {
            "description": "Private Channel은 수정할 수 없음",
            "content": {
              "*/*": {
                "example": "Private channel cannot be updated"
              }
            }
          },
          "200": {
            "description": "Channel 정보가 성공적으로 수정됨",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/channels": {
      "get": {
        "tags": [
          "Channel"
        ],
        "summary": "User가 참여 중인 Channel 목록 조회",
        "operationId": "findAll_1",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "조회할 User ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Channel 목록 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChannelDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BinaryContentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fileName": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "contentType": {
            "type": "string"
          }
        }
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "profile": {
            "$ref": "#/components/schemas/BinaryContentDto"
          },
          "online": {
            "type": "boolean"
          }
        }
      },
      "ReadStatusCreateRequest": {
        "type": "object",
        "description": "Message 읽음 상태 생성 정보",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "channelId": {
            "type": "string",
            "format": "uuid"
          },
          "lastReadAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ReadStatusDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "channelId": {
            "type": "string",
            "format": "uuid"
          },
          "lastReadAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MessageCreateRequest": {
        "type": "object",
        "description": "Message 생성 정보",
        "properties": {
          "content": {
            "type": "string"
          },
          "channelId": {
            "type": "string",
            "format": "uuid"
          },
          "authorId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "MessageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "content": {
            "type": "string"
          },
          "channelId": {
            "type": "string",
            "format": "uuid"
          },
          "author": {
            "$ref": "#/components/schemas/UserDto"
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BinaryContentDto"
            }
          }
        }
      },
      "PublicChannelCreateRequest": {
        "type": "object",
        "description": "Public Channel 생성 정보",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "ChannelDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ]
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserDto"
            }
          },
          "lastMessageAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PrivateChannelCreateRequest": {
        "type": "object",
        "description": "Private Channel 생성 정보",
        "properties": {
          "participantIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "MessageUpdateRequest": {
        "type": "object",
        "description": "수정할 Message 내용",
        "properties": {
          "newContent": {
            "type": "string"
          }
        }
      },
      "PublicChannelUpdateRequest": {
        "type": "object",
        "description": "수정할 Channel 정보",
        "properties": {
          "newName": {
            "type": "string"
          },
          "newDescription": {
            "type": "string"
          }
        }
      },
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "PageResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "nextCursor": {
            "type": "object"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "hasNext": {
            "type": "boolean"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          }
        }
      }
    }
  }
}