# src/main/java/com/sprint/mission/discodeit/controller/ChannelController.java
package com.sprint.mission.discodeit.controller;

import com.sprint.mission.discodeit.controller.api.ChannelApi;
import com.sprint.mission.discodeit.dto.data.ChannelDto;
import com.sprint.mission.discodeit.dto.request.PrivateChannelCreateRequest;
import com.sprint.mission.discodeit.dto.request.PublicChannelCreateRequest;
import com.sprint.mission.discodeit.dto.request.PublicChannelUpdateRequest;
import com.sprint.mission.discodeit.security.CustomUserDetails;
import com.sprint.mission.discodeit.service.ChannelService;
import jakarta.validation.Valid;
import java.util.List;
import java.util.UUID;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.security.core.Authentication;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PatchMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;
import lombok.extern.slf4j.Slf4j;

@Slf4j
@RequiredArgsConstructor
@RestController
@RequestMapping("/api/channels")
public class ChannelController implements ChannelApi {

  private final ChannelService channelService;

  @PreAuthorize("hasRole('ROLE_CHANNEL_MANAGER')")
  @PostMapping(path = "public")
  public ResponseEntity<ChannelDto> create(@RequestBody @Valid PublicChannelCreateRequest request) {
    log.info("공개 채널 생성 요청: {}", request);
    ChannelDto createdChannel = channelService.create(request);
    log.debug("공개 채널 생성 응답: {}", createdChannel);
    return ResponseEntity
        .status(HttpStatus.CREATED)
        .body(createdChannel);
  }

  @PostMapping(path = "private")
  public ResponseEntity<ChannelDto> create(
      @RequestBody @Valid PrivateChannelCreateRequest request) {
    log.info("비공개 채널 생성 요청: {}", request);
    ChannelDto createdChannel = channelService.create(request);
    log.debug("비공개 채널 생성 응답: {}", createdChannel);
    return ResponseEntity
        .status(HttpStatus.CREATED)
        .body(createdChannel);
  }

  @PreAuthorize("hasRole('ROLE_CHANNEL_MANAGER')")
  @PatchMapping(path = "{channelId}")
  public ResponseEntity<ChannelDto> update(
      @PathVariable("channelId") UUID channelId,
      @RequestBody @Valid PublicChannelUpdateRequest request) {
    log.info("채널 수정 요청: id={}, request={}", channelId, request);
    ChannelDto updatedChannel = channelService.update(channelId, request);
    log.debug("채널 수정 응답: {}", updatedChannel);
    return ResponseEntity
        .status(HttpStatus.OK)
        .body(updatedChannel);
  }

  @PreAuthorize("hasRole('ROLE_CHANNEL_MANAGER')")
  @DeleteMapping(path = "{channelId}")
  public ResponseEntity<Void> delete(@PathVariable("channelId") UUID channelId) {
    log.info("채널 삭제 요청: id={}", channelId);
    channelService.delete(channelId);
    log.debug("채널 삭제 완료");
    return ResponseEntity
        .status(HttpStatus.NO_CONTENT)
        .build();
  }

  @GetMapping
  public ResponseEntity<List<ChannelDto>> findAll(
      @RequestParam(required = false) String userId,
      Authentication authentication) {

    UUID actualUserId;
    
    if (userId == null || "undefined".equals(userId) || userId.trim().isEmpty()) {
      // 현재 인증된 사용자 사용
      log.debug("userId가 null 또는 undefined이므로 현재 사용자 ID 사용");
      CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
      actualUserId = userDetails.getUser().getId();
    } else {
      try {
        actualUserId = UUID.fromString(userId);
        log.debug("요청된 userId 사용: {}", actualUserId);
      } catch (IllegalArgumentException e) {
        // 유효하지 않은 UUID인 경우 현재 사용자 사용
        log.warn("유효하지 않은 UUID 형식: {}. 현재 사용자 ID 사용", userId);
        CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
        actualUserId = userDetails.getUser().getId();
      }
    }

    List<ChannelDto> channels = channelService.findAllByUserId(actualUserId);
    return ResponseEntity.ok(channels);
  }
}